wlan: Don't disconnect in case of HS2.0 for change in AP capability.
After connection, driver process beacon  to detect any change in
AP's capabilities. Currently if privacy bit is set in beacon and
if rsn or wpa bit is not set, driver issues disconnect command.
In case of HS2.0, all security capabilities are part of vendor
specific information. Hence, both wpa and rsn bit is 0, though
privacy bit is set. Since driver considers it as a change in AP
security params, it issues disconnect. As a part of fix, make sure
that for OSEN capabilities disconnect is not issued.

Change-Id: I5554b24dfa2724c6c9b6faef2f738277e42028fb
CRs-Fixed: 796397