wlan: Reset LoadUnloadInProgress flag in FTM mode.
Currently in FTM mode, pHddCtx->isLoadUnloadInProgress is set
during the driver load and it is not reset once the driver is
loaded successfully.Because of this any ioctl issued will be
failed as ioctl has check for load/unload in progress.

To resolve this issue, isLoadUnloadInProgress flag is reset once
the driver is loaded successfully in FTM mode.

Change-Id: Ic01506dd313dd2b70604acae5de85e0808dd54f3
CRs-Fixed: 806556